O3-4865: Modifying the VisitPatientScenario
