+++ src/test/java/org/openmrs/performance/http/ClerkHttpService.java
@@ -20,6 +20,7 @@ import static io.gatling.javaapi.http.HttpDsl.http;
 import static org.openmrs.performance.Constants.ADMIN_SUPER_USER_UUID;
 import static org.openmrs.performance.Constants.OUTPATIENT_CLINIC_LOCATION_UUID;
 import static org.openmrs.performance.Constants.GENERAL_MEDICINE_SERVICE_UUID;
+import static org.openmrs.performance.Constants.PATIENT_IDENTIFIER_UUID;
 import static org.openmrs.performance.Constants.USER_GENERATED_PATIENT_LIST;
 import static org.openmrs.performance.utils.CommonUtils.getAdjustedDateTimeAsString;
 import static org.openmrs.performance.utils.CommonUtils.getCurrentDateTimeAsString;
@@ -126,9 +127,9 @@ public class ClerkHttpService extends HttpService {
 
 	public HttpRequestActionBuilder getPatientQueueEntry(String patientUuid) {
 		String customRepresentation = "custom:(uuid,display,queue,status,patient:(uuid,display,person,identifiers:(uuid,display,identifier,identifierType)),"
-				+ "visit:(uuid,display,startDatetime,encounters:(uuid,display,diagnoses,encounterDatetime,encounterType,obs,encounterProviders,voided),"
-				+ "attributes:(uuid,display,value,attributeType)),priority,priorityComment,sortWeight,startedAt,endedAt,locationWaitingFor,queueComingFrom,"
-				+ "providerWaitingFor,previousQueueEntry)";
+		        + "visit:(uuid,display,startDatetime,encounters:(uuid,display,diagnoses,encounterDatetime,encounterType,obs,encounterProviders,voided),"
+		        + "attributes:(uuid,display,value,attributeType)),priority,priorityComment,sortWeight,startedAt,endedAt,locationWaitingFor,queueComingFrom,"
+		        + "providerWaitingFor,previousQueueEntry)";
 
 		return http("Get Queue Entry").get("/openmrs/ws/rest/v1/queue-entry?v=" + customRepresentation
 		        + "&totalCount=true&patient=" + patientUuid + "&isEnded=false");
@@ -258,5 +259,80 @@ public class ClerkHttpService extends HttpService {
 	public HttpRequestActionBuilder getMembersOfPatientList(String patientListUuid) {
 		return http("Fetch members of patient list").get(
 		    "/openmrs/ws/rest/v1/cohortm/cohortmember?cohort=" + patientListUuid + "&startIndex=0&limit=10&v=full&q=");
+	}
+
+	public HttpRequestActionBuilder getPatientAttributes(String personUuid) {
+		String customRepresentation = "custom:(uuid,display,attributeType:(uuid,display,format),value)";
+		return http("Get Person Attributes")
+		        .get("/openmrs/ws/rest/v1/person/" + personUuid + "/attribute?v=" + customRepresentation);
+	}
+
+	public HttpRequestActionBuilder getPatientIdentifiers(String patientUuid) {
+		String customRepresentation = "custom:(uuid,identifier,identifierType:(uuid,required,name),preferred)";
+		return http("Get Patient Identifiers")
+		        .get("/openmrs/ws/rest/v1/patient/" + patientUuid + "/identifier?v=" + customRepresentation);
+	}
+
+	public HttpRequestActionBuilder getPatientRelationships(String personUuid) {
+		String customRepresentation = "custom:(display,uuid,personA:(age,display,birthdate,uuid),personB:(age,display,birthdate,uuid),relationshipType:(uuid,display,description,aIsToB,bIsToA))";
+		return http("Get Relationships of Person")
+		        .get("/openmrs/ws/rest/v1/relationship?v=" + customRepresentation + "&person=" + personUuid);
+	}
+
+	public HttpRequestActionBuilder editPatientDetails(String patientUuid) {
+		return http("Edit patient details").post("/openmrs/ws/rest/v1/patient/" + patientUuid).body(StringBody(session -> {
+			Map<String, Object> payload = new HashMap<>();
+
+			payload.put("uuid", session.get("patient_uuid"));
+
+			Map<String, Object> person = new HashMap<>();
+			person.put("uuid", session.get("patient_uuid"));
+
+			List<Map<String, Object>> names = new ArrayList<>();
+			Map<String, Object> nameItem = new HashMap<>();
+			nameItem.put("uuid", session.getString("patientNameId"));
+			nameItem.put("preferred", true);
+			nameItem.put("givenName", "Mark");
+			nameItem.put("familyName", "Williams");
+			names.add(nameItem);
+			person.put("names", names);
+
+			person.put("gender", "M");
+			person.put("birthdate", "1962-4-5");
+			person.put("birthdateEstimated", false);
+			person.put("attributes", new ArrayList<>());
+
+			List<Map<String, Object>> addresses = new ArrayList<>();
+			Map<String, Object> address = new HashMap<>();
+			address.put("address1", "Address16582");
+			address.put("cityVillage", "City6582");
+			address.put("stateProvince", "State6582");
+			address.put("postalCode", "898989");
+			address.put("country", "Country6582");
+			addresses.add(address);
+			person.put("addresses", addresses);
+
+			person.put("dead", false);
+
+			payload.put("person", person);
+
+			List<Map<String, Object>> identifiers = new ArrayList<>();
+			Map<String, Object> identifier = new HashMap<>();
+			identifier.put("uuid", session.getString("patientIdentifierId"));
+			identifier.put("identifier", session.getString("patientIdentifierValue"));
+			identifier.put("identifierType", PATIENT_IDENTIFIER_UUID);
+			identifier.put("location", OUTPATIENT_CLINIC_LOCATION_UUID);
+			identifier.put("preferred", true);
+			identifiers.add(identifier);
+
+			payload.put("identifiers", identifiers);
+
+			try {
+				return new ObjectMapper().writeValueAsString(payload);
+			}
+			catch (JsonProcessingException e) {
+				throw new RuntimeException("Error converting identifiers to JSON", e);
+			}
+		}));
 	}
 }
