+++ src/test/java/org/openmrs/performance/http/DoctorHttpService.java
@@ -86,54 +86,6 @@ public class DoctorHttpService extends HttpService {
 		        + customRepresentation + "&excludeDiscontinueOrders=true");
 	}
 
-	public HttpRequestActionBuilder getAllergies(String patientUuid) {
-		return http("Get Allergies of Patient")
-		        .get("/openmrs/ws/fhir2/R4/AllergyIntolerance?patient=" + patientUuid + "&_summary=data");
-	}
-
-	public HttpRequestActionBuilder getAllergens(String allergenType, String allergenUuid) {
-		return http("Get " + allergenType + " Allergens").get("/openmrs/ws/rest/v1/concept/" + allergenUuid + "?v=full");
-	}
-
-	public HttpRequestActionBuilder saveAllergy(String patientUuid) {
-		// Using the 'OTHER' allergen type to create a unique entry and avoid duplication
-		return http("Save an Allergy").post("/openmrs/ws/rest/v1/patient/" + patientUuid + "/allergy")
-		        .body(StringBody(session -> {
-			        try {
-				        String random = new Random().ints(7, 'a', 'z' + 1)
-				                .collect(StringBuilder::new, StringBuilder::append, StringBuilder::append).toString();
-				        Map<String, Object> payload = new HashMap<>();
-				        Map<String, String> codedAllergen = new HashMap<>();
-
-				        codedAllergen.put("uuid", OTHER_NON_CODED_ALLERGEN_UUID);
-				        Map<String, Object> allergen = new HashMap<>();
-
-				        allergen.put("allergenType", "OTHER");
-				        allergen.put("codedAllergen", codedAllergen);
-				        allergen.put("nonCodedAllergen", random);
-
-				        Map<String, String> severity = new HashMap<>();
-				        severity.put("uuid", SEVERITY_UUID);
-
-				        Map<String, String> reactionUuid = new HashMap<>();
-				        reactionUuid.put("uuid", ALLERGY_REACTION_UUID);
-
-				        Map<String, Object> reaction = new HashMap<>();
-				        reaction.put("reaction", reactionUuid);
-
-				        List<Map<String, Object>> reactions = Collections.singletonList(reaction);
-				        payload.put("allergen", allergen);
-				        payload.put("severity", severity);
-				        payload.put("comment", "test");
-				        payload.put("reactions", reactions);
-				        return new ObjectMapper().writeValueAsString(payload);
-			        }
-			        catch (JsonProcessingException e) {
-				        throw new RuntimeException(e);
-			        }
-		        }));
-	}
-
 	public HttpRequestActionBuilder getAttachments(String patientUuid) {
 		return http("Get Attachments of Patient")
 		        .get("/openmrs/ws/rest/v1/attachment?patient=" + patientUuid + "&includeEncounterless=true");
@@ -259,113 +211,6 @@ public class DoctorHttpService extends HttpService {
 		catch (JsonProcessingException e) {
 			throw new RuntimeException("Error converting patientDiagnosis to JSON", e);
 		}
-	}
-
-	public HttpRequestActionBuilder saveVitalsData(String patientUuid) {
-		String encounterDatetime = CommonUtils.getCurrentDateTimeAsString();
-
-		Map<String, Object> encounter = new HashMap<>();
-		encounter.put("form", VITALS_FORM_UUID);
-		encounter.put("patient", patientUuid);
-		encounter.put("location", VITALS_LOCATION_UUID);
-		encounter.put("encounterType", VITALS_ENCOUNTER_TYPE_UUID);
-		encounter.put("encounterDatetime", encounterDatetime);
-
-		List<Map<String, Object>> observations = new ArrayList<>();
-		observations.add(Map.of("concept", SYSTOLIC_BLOOD_PRESSURE, "value", 34));
-		observations.add(Map.of("concept", DIASTOLIC_BLOOD_PRESSURE, "value", 44));
-		observations.add(Map.of("concept", RESPIRATORY_RATE, "value", 20));
-		observations.add(Map.of("concept", ARTERIAL_BLOOD_OXYGEN_SATURATION, "value", 20));
-		observations.add(Map.of("concept", PULSE, "value", 120));
-		observations.add(Map.of("concept", TEMPERATURE_C, "value", 28));
-		observations.add(Map.of("concept", WEIGHT_KG, "value", 60));
-		observations.add(Map.of("concept", HEIGHT_CM, "value", 121));
-		observations.add(Map.of("concept", MID_UPPER_ARM_CIRCUMFERENCE, "value", 34));
-
-		encounter.put("obs", observations);
-
-		try {
-			String body = new ObjectMapper().writeValueAsString(encounter); // Convert Map to JSON
-			return http("Save Vitals").post("/openmrs/ws/rest/v1/encounter").body(StringBody(body));
-		}
-		catch (JsonProcessingException e) {
-			throw new RuntimeException("Error converting visitNote to JSON", e);
-		}
-	}
-
-	public HttpRequestActionBuilder getPatientAttributes(String personUuid) {
-		String customRepresentation = "custom:(uuid,display,attributeType:(uuid,display,format),value)";
-		return http("Get Person Attributes")
-		        .get("/openmrs/ws/rest/v1/person/" + personUuid + "/attribute?v=" + customRepresentation);
-	}
-
-	public HttpRequestActionBuilder getPatientIdentifiers(String patientUuid) {
-		String customRepresentation = "custom:(uuid,identifier,identifierType:(uuid,required,name),preferred)";
-		return http("Get Patient Identifiers")
-		        .get("/openmrs/ws/rest/v1/patient/" + patientUuid + "/identifier?v=" + customRepresentation);
-	}
-
-	public HttpRequestActionBuilder getPatientRelationships(String personUuid) {
-		String customRepresentation = "custom:(display,uuid,personA:(age,display,birthdate,uuid),personB:(age,display,birthdate,uuid),relationshipType:(uuid,display,description,aIsToB,bIsToA))";
-		return http("Get Relationships of Person")
-		        .get("/openmrs/ws/rest/v1/relationship?v=" + customRepresentation + "&person=" + personUuid);
-	}
-
-	public HttpRequestActionBuilder editPatientDetails(String patientUuid) {
-		return http("Edit patient details").post("/openmrs/ws/rest/v1/patient/" + patientUuid).body(StringBody(session -> {
-			Map<String, Object> payload = new HashMap<>();
-
-			payload.put("uuid", session.get("patient_uuid"));
-
-			Map<String, Object> person = new HashMap<>();
-			person.put("uuid", session.get("patient_uuid"));
-
-			List<Map<String, Object>> names = new ArrayList<>();
-			Map<String, Object> nameItem = new HashMap<>();
-			nameItem.put("uuid", session.getString("patientNameId"));
-			nameItem.put("preferred", true);
-			nameItem.put("givenName", "Mark");
-			nameItem.put("familyName", "Williams");
-			names.add(nameItem);
-			person.put("names", names);
-
-			person.put("gender", "M");
-			person.put("birthdate", "1962-4-5");
-			person.put("birthdateEstimated", false);
-			person.put("attributes", new ArrayList<>());
-
-			List<Map<String, Object>> addresses = new ArrayList<>();
-			Map<String, Object> address = new HashMap<>();
-			address.put("address1", "Address16582");
-			address.put("cityVillage", "City6582");
-			address.put("stateProvince", "State6582");
-			address.put("postalCode", "898989");
-			address.put("country", "Country6582");
-			addresses.add(address);
-			person.put("addresses", addresses);
-
-			person.put("dead", false);
-
-			payload.put("person", person);
-
-			List<Map<String, Object>> identifiers = new ArrayList<>();
-			Map<String, Object> identifier = new HashMap<>();
-			identifier.put("uuid", session.getString("patientIdentifierId"));
-			identifier.put("identifier", session.getString("patientIdentifierValue"));
-			identifier.put("identifierType", PATIENT_IDENTIFIER_UUID);
-			identifier.put("location", OUTPATIENT_CLINIC_LOCATION_UUID);
-			identifier.put("preferred", true);
-			identifiers.add(identifier);
-
-			payload.put("identifiers", identifiers);
-
-			try {
-				return new ObjectMapper().writeValueAsString(payload);
-			}
-			catch (JsonProcessingException e) {
-				throw new RuntimeException("Error converting identifiers to JSON", e);
-			}
-		}));
 	}
 
 	public HttpRequestActionBuilder addProgramEnrollment() {
