+++ src/test/java/org/openmrs/performance/http/NurseHttpService.java
@@ -3,22 +3,41 @@ package org.openmrs.performance.http;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.gatling.javaapi.http.HttpRequestActionBuilder;
+import org.openmrs.performance.utils.CommonUtils;
 
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Random;
 import java.util.Set;
 import java.util.StringJoiner;
 
 import static io.gatling.javaapi.core.CoreDsl.StringBody;
 import static io.gatling.javaapi.core.CoreDsl.jsonPath;
 import static io.gatling.javaapi.http.HttpDsl.http;
+import static org.openmrs.performance.Constants.ALLERGY_REACTION_UUID;
+import static org.openmrs.performance.Constants.ARTERIAL_BLOOD_OXYGEN_SATURATION;
 import static org.openmrs.performance.Constants.BED_ASSIGNMENT_UUID;
 import static org.openmrs.performance.Constants.CLINICIAN_ENCOUNTER_ROLE;
+import static org.openmrs.performance.Constants.DIASTOLIC_BLOOD_PRESSURE;
+import static org.openmrs.performance.Constants.HEIGHT_CM;
+import static org.openmrs.performance.Constants.MID_UPPER_ARM_CIRCUMFERENCE;
+import static org.openmrs.performance.Constants.OTHER_NON_CODED_ALLERGEN_UUID;
+import static org.openmrs.performance.Constants.PULSE;
+import static org.openmrs.performance.Constants.RESPIRATORY_RATE;
+import static org.openmrs.performance.Constants.SEVERITY_UUID;
+import static org.openmrs.performance.Constants.SYSTOLIC_BLOOD_PRESSURE;
+import static org.openmrs.performance.Constants.TEMPERATURE_C;
 import static org.openmrs.performance.Constants.TEST_ORDER_TYPE_UUID;
+import static org.openmrs.performance.Constants.VITALS_ENCOUNTER_TYPE_UUID;
+import static org.openmrs.performance.Constants.VITALS_FORM_UUID;
+import static org.openmrs.performance.Constants.VITALS_LOCATION_UUID;
 import static org.openmrs.performance.Constants.WARD_ADMISSION_NOTE_UUID;
+import static org.openmrs.performance.Constants.WEIGHT_KG;
 import static org.openmrs.performance.utils.CommonUtils.getCurrentDateTimeAsString;
 
 public class NurseHttpService extends HttpService {
@@ -160,6 +179,86 @@ public class NurseHttpService extends HttpService {
 
 		return http("Get patient admission note").get("/openmrs/ws/rest/v1/obs?patient=" + patientUuid + "&concept="
 		        + WARD_ADMISSION_NOTE_UUID + "&v=" + customRepresentation);
+	}
+
+	public HttpRequestActionBuilder getAllergies(String patientUuid) {
+		return http("Get Allergies of Patient")
+		        .get("/openmrs/ws/fhir2/R4/AllergyIntolerance?patient=" + patientUuid + "&_summary=data");
+	}
+
+	public HttpRequestActionBuilder getAllergens(String allergenType, String allergenUuid) {
+		return http("Get " + allergenType + " Allergens").get("/openmrs/ws/rest/v1/concept/" + allergenUuid + "?v=full");
+	}
+
+	public HttpRequestActionBuilder saveAllergy(String patientUuid) {
+		// Using the 'OTHER' allergen type to create a unique entry and avoid duplication
+		return http("Save an Allergy").post("/openmrs/ws/rest/v1/patient/" + patientUuid + "/allergy")
+		        .body(StringBody(session -> {
+			        try {
+				        String random = new Random().ints(7, 'a', 'z' + 1)
+				                .collect(StringBuilder::new, StringBuilder::append, StringBuilder::append).toString();
+				        Map<String, Object> payload = new HashMap<>();
+				        Map<String, String> codedAllergen = new HashMap<>();
+
+				        codedAllergen.put("uuid", OTHER_NON_CODED_ALLERGEN_UUID);
+				        Map<String, Object> allergen = new HashMap<>();
+
+				        allergen.put("allergenType", "OTHER");
+				        allergen.put("codedAllergen", codedAllergen);
+				        allergen.put("nonCodedAllergen", random);
+
+				        Map<String, String> severity = new HashMap<>();
+				        severity.put("uuid", SEVERITY_UUID);
+
+				        Map<String, String> reactionUuid = new HashMap<>();
+				        reactionUuid.put("uuid", ALLERGY_REACTION_UUID);
+
+				        Map<String, Object> reaction = new HashMap<>();
+				        reaction.put("reaction", reactionUuid);
+
+				        List<Map<String, Object>> reactions = Collections.singletonList(reaction);
+				        payload.put("allergen", allergen);
+				        payload.put("severity", severity);
+				        payload.put("comment", "test");
+				        payload.put("reactions", reactions);
+				        return new ObjectMapper().writeValueAsString(payload);
+			        }
+			        catch (JsonProcessingException e) {
+				        throw new RuntimeException(e);
+			        }
+		        }));
+	}
+
+	public HttpRequestActionBuilder saveVitalsData(String patientUuid) {
+		String encounterDatetime = CommonUtils.getCurrentDateTimeAsString();
+
+		Map<String, Object> encounter = new HashMap<>();
+		encounter.put("form", VITALS_FORM_UUID);
+		encounter.put("patient", patientUuid);
+		encounter.put("location", VITALS_LOCATION_UUID);
+		encounter.put("encounterType", VITALS_ENCOUNTER_TYPE_UUID);
+		encounter.put("encounterDatetime", encounterDatetime);
+
+		List<Map<String, Object>> observations = new ArrayList<>();
+		observations.add(Map.of("concept", SYSTOLIC_BLOOD_PRESSURE, "value", 34));
+		observations.add(Map.of("concept", DIASTOLIC_BLOOD_PRESSURE, "value", 44));
+		observations.add(Map.of("concept", RESPIRATORY_RATE, "value", 20));
+		observations.add(Map.of("concept", ARTERIAL_BLOOD_OXYGEN_SATURATION, "value", 20));
+		observations.add(Map.of("concept", PULSE, "value", 120));
+		observations.add(Map.of("concept", TEMPERATURE_C, "value", 28));
+		observations.add(Map.of("concept", WEIGHT_KG, "value", 60));
+		observations.add(Map.of("concept", HEIGHT_CM, "value", 121));
+		observations.add(Map.of("concept", MID_UPPER_ARM_CIRCUMFERENCE, "value", 34));
+
+		encounter.put("obs", observations);
+
+		try {
+			String body = new ObjectMapper().writeValueAsString(encounter); // Convert Map to JSON
+			return http("Save Vitals").post("/openmrs/ws/rest/v1/encounter").body(StringBody(body));
+		}
+		catch (JsonProcessingException e) {
+			throw new RuntimeException("Error converting visitNote to JSON", e);
+		}
 	}
 
 }
