+++ src/test/java/org/openmrs/performance/personas/ClerkPersona.java
@@ -3,6 +3,7 @@ package org.openmrs.performance.personas;
 import org.openmrs.performance.registries.ClerkRegistry;
 import org.openmrs.performance.scenarios.PatientAppointmentCreationScenario;
 import org.openmrs.performance.scenarios.PatientListCreationScenario;
+import org.openmrs.performance.scenarios.PatientProfileUpdateScenario;
 import org.openmrs.performance.scenarios.PatientRegistrationScenario;
 import org.openmrs.performance.scenarios.Scenario;
 
@@ -16,7 +17,7 @@ public class ClerkPersona extends Persona<ClerkRegistry> {
 
 	@Override
 	public List<Scenario<ClerkRegistry>> getScenarios() {
-		return List.of(new PatientRegistrationScenario(0.4F), new PatientAppointmentCreationScenario(0.4F),
-		    new PatientListCreationScenario(0.2F));
+		return List.of(new PatientRegistrationScenario(0.3F), new PatientAppointmentCreationScenario(0.3F),
+		    new PatientListCreationScenario(0.2F), new PatientProfileUpdateScenario(0.2F));
 	}
 }
