+++ src/test/java/org/openmrs/performance/personas/NursePersona.java
@@ -1,6 +1,7 @@
 package org.openmrs.performance.personas;
 
 import org.openmrs.performance.registries.NurseRegistry;
+import org.openmrs.performance.scenarios.PatientInitialAssessmentScenario;
 import org.openmrs.performance.scenarios.PatientWardAdmissionScenario;
 import org.openmrs.performance.scenarios.Scenario;
 
@@ -14,7 +15,7 @@ public class NursePersona extends Persona<NurseRegistry> {
 
 	@Override
 	public List<Scenario<NurseRegistry>> getScenarios() {
-		return List.of(new PatientWardAdmissionScenario(1));
+		return List.of(new PatientWardAdmissionScenario(0.5F), new PatientInitialAssessmentScenario(0.5F));
 	}
 
 }
