+++ src/test/java/org/openmrs/performance/scenarios/VisitPatientScenario.java
@@ -35,26 +35,12 @@ public class VisitPatientScenario extends Scenario<DoctorRegistry> {
 				.exec(registry.saveSoapTemplateForm("#{patient_uuid}"))
 				.pause(3)
 				.exec(registry.openVisitsTab("#{patient_uuid}"))
-				.pause(5)
-				.exec(registry.openEditPatientTab("#{patient_uuid}"))
-				.pause(5)
-				.exec(registry.editPatientDetails("#{patient_uuid}"))
 				.pause(2)
 				.exec(registry.getVisitsFromNewEndpoint("#{patient_uuid}"))
-				.pause(5)
-				.exec(registry.openVitalsAndBiometricsTab("#{patient_uuid}"))
-				.pause(5)
-				.exec(registry.recordVitals("#{patient_uuid}"))
 				.pause(5)
 				.exec(registry.openMedicationsTab("#{patient_uuid}"))
 				.pause(5)
 				.exec(registry.openLabResultsTab("#{patient_uuid}"))
-				.pause(8)
-				.exec(registry.openAllergiesTab("#{patient_uuid}"))
-				.pause(5)
-				.exec(registry.openAllergiesForm())
-				.pause(5)
-				.exec(registry.recordAllergy("#{patient_uuid}"))
 				.pause(10)
 				.exec(registry.openConditionsTab("#{patient_uuid}"))
 				.pause(5)
